Move db context to data access
