+++ PollsAppBlazor/Server/Program.cs
@@ -1,5 +1,5 @@
 using Microsoft.EntityFrameworkCore;
-using PollsAppBlazor.Server.Data;
+using PollsAppBlazor.Server.DataAccess;
 using PollsAppBlazor.Server.Extensions;
 using PollsAppBlazor.Server.Services;
 using System.Text.Json.Serialization;
@@ -9,7 +9,7 @@ var builder = WebApplication.CreateBuilder(args);
 // Add services to the container.
 var connectionString = builder.Configuration.GetConnectionString("DefaultConnection") ?? throw new InvalidOperationException("Connection string 'DefaultConnection' not found.");
 builder.Services.AddDbContext<ApplicationDbContext>(options =>
-	options.UseSqlServer(connectionString));
+    options.UseSqlServer(connectionString));
 
 builder.Services.AddScoped<PollsService>();
 builder.Services.AddScoped<OptionsService>();
@@ -24,12 +24,12 @@ builder.Services.AddCustomizedIdentity();
 builder.Services.AddCustomizedAuthorization();
 
 builder.Services
-	.AddControllers()
-	.AddJsonOptions(options =>
-	{
-		options.JsonSerializerOptions.DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull;
-	})
-	.ConfigureApiBehaviorOptions(x => { x.SuppressMapClientErrors = true; });
+    .AddControllers()
+    .AddJsonOptions(options =>
+    {
+        options.JsonSerializerOptions.DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull;
+    })
+    .ConfigureApiBehaviorOptions(x => { x.SuppressMapClientErrors = true; });
 builder.Services.AddRazorPages();
 
 builder.Services.AddSwagger();
@@ -39,18 +39,18 @@ var app = builder.Build();
 // Configure the HTTP request pipeline.
 if (app.Environment.IsDevelopment())
 {
-	app.UseSwagger();
-	app.UseSwaggerUI();
+    app.UseSwagger();
+    app.UseSwaggerUI();
 
 
-	app.UseMigrationsEndPoint();
-	app.UseWebAssemblyDebugging();
+    app.UseMigrationsEndPoint();
+    app.UseWebAssemblyDebugging();
 }
 else
 {
-	app.UseExceptionHandler("/Error");
-	// The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-	app.UseHsts();
+    app.UseExceptionHandler("/Error");
+    // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+    app.UseHsts();
 }
 
 app.UseHttpsRedirection();
